requirements.txt
django-elasticsearch-dsl
channels
djangorestframework
django-cors-headers